rename frauddetectionservice to fraud-detection
