+++ .env
@@ -74,7 +74,7 @@ EMAIL_SERVICE_ADDR=http://emailservice:${EMAIL_SERVICE_PORT}
 EMAIL_SERVICE_DOCKERFILE=./src/emailservice/Dockerfile
 
 # Fraud Service
-FRAUD_SERVICE_DOCKERFILE=./src/frauddetectionservice/Dockerfile
+FRAUD_DOCKERFILE=./src/fraud-detection/Dockerfile
 
 # Frontend
 FRONTEND_PORT=8080
