+++ .github/workflows/component-build-images.yml
@@ -75,8 +75,8 @@ jobs:
             tag_suffix: emailservice
             context: ./
             setup-qemu: true
-          - file: ./src/frauddetectionservice/Dockerfile
-            tag_suffix: frauddetectionservice
+          - file: ./src/fraud-detection/Dockerfile
+            tag_suffix: fraud-detection
             context: ./
             setup-qemu: true
           - file: ./src/frontend/Dockerfile
