+++ .gitignore
@@ -49,7 +49,7 @@ test/tracetesting/tracetesting-vars.yaml
 /src/featureflagservice/src/ffs_demo_pb.erl
 /src/featureflagservice/src/ffs_service_*.erl
 /src/featureflagservice/src/oteldemo_*.erl
-/src/frauddetectionservice/src/main/proto
+/src/fraud-detection/src/main/proto
 /src/frontend/pb/
 /src/frontend/protos/
 /src/paymentservice/demo.proto
