+++ docker-compose-tests.yml
@@ -69,7 +69,7 @@ services:
         condition: service_started
       emailservice:
         condition: service_started
-      frauddetectionservice:
+      fraud-detection:
         condition: service_started
       frontend:
         condition: service_started
