+++ src/fraud-detection/Dockerfile
@@ -6,7 +6,7 @@ FROM --platform=${BUILDPLATFORM} gradle:8-jdk17 AS builder
 
 WORKDIR /usr/src/app/
 
-COPY ./src/frauddetectionservice/ ./
+COPY ./src/fraud-detection/ ./
 COPY ./pb/ ./src/main/proto/
 RUN gradle shadowJar
 
@@ -17,8 +17,8 @@ FROM gcr.io/distroless/java17-debian11
 ARG OTEL_JAVA_AGENT_VERSION
 WORKDIR /usr/src/app/
 
-COPY --from=builder /usr/src/app/build/libs/frauddetectionservice-1.0-all.jar ./
+COPY --from=builder /usr/src/app/build/libs/fraud-detection-1.0-all.jar ./
 ADD --chmod=644 https://github.com/open-telemetry/opentelemetry-java-instrumentation/releases/download/v$OTEL_JAVA_AGENT_VERSION/opentelemetry-javaagent.jar /app/opentelemetry-javaagent.jar
 ENV JAVA_TOOL_OPTIONS=-javaagent:/app/opentelemetry-javaagent.jar
 
-ENTRYPOINT [ "java", "-jar", "frauddetectionservice-1.0-all.jar" ]
+ENTRYPOINT [ "java", "-jar", "fraud-detection-1.0-all.jar" ]
