+++ src/fraud-detection/gradle.properties
