--- src/frauddetectionservice/settings.gradle.kts
@@ -1,3 +1,0 @@
-
-rootProject.name = "frauddetectionservice"
-
